Remove devdocs from System.Runtime.InteropServices code (#36669)
* Cleanup System.Runtime.InteropServices code

- Move public classes/enums to their own file
- Use summary instead of devdoc
- Misc
