+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/AutomationProxyAttribute.cs
@@ -1,0 +1,14 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Class | AttributeTargets.Interface, Inherited = false)]
+    public sealed class AutomationProxyAttribute : Attribute
+    {
+        public AutomationProxyAttribute(bool val) => Value = val;
+
+        public bool Value { get; }
+    }
+}
