+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/ComConversionLossAttribute.cs
@@ -4,10 +4,8 @@
 
 namespace System.Runtime.InteropServices
 {
-    [Flags]
-    public enum AssemblyRegistrationFlags
+    [AttributeUsage(AttributeTargets.All, Inherited = false)]
+    public sealed class ComConversionLossAttribute : Attribute
     {
-        None = 0x00000000,
-        SetCodeBase = 0x00000001,
     }
 }
