+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/ComTypes/ADVF.cs
@@ -4,6 +4,11 @@
 
 namespace System.Runtime.InteropServices.ComTypes
 {
+    /// <remarks>
+    /// Note: ADVF_ONLYONCE and ADVF_PRIMEFIRST values conform with objidl.dll but are backwards from 
+    /// the Platform SDK documentation as of 07/21/2003.
+    /// https://docs.microsoft.com/en-us/windows/desktop/api/objidl/ne-objidl-tagadvf.
+    /// </remarks>
     [Flags]
     public enum ADVF
     {
@@ -15,7 +20,4 @@ namespace System.Runtime.InteropServices.ComTypes
         ADVFCACHE_FORCEBUILTIN = 16,
         ADVFCACHE_ONSAVE = 32
     }
-    // Note: ADVF_ONLYONCE and ADVF_PRIMEFIRST values conform with objidl.dll but are backwards from 
-    // the Platform SDK documentation as of 07/21/2003.
-    // https://docs.microsoft.com/en-us/windows/desktop/api/objidl/ne-objidl-tagadvf.
 }
