+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/ComTypes/IDataObject.cs
@@ -4,7 +4,7 @@
 
 namespace System.Runtime.InteropServices.ComTypes
 {
-    /// <devdoc>
+    /// <summary>
     ///     The IDataObject interface specifies methods that enable data transfer
     ///     and notification of changes in data. Data transfer methods specify
     ///     the format of the transferred data along with the medium through
@@ -13,75 +13,75 @@ namespace System.Runtime.InteropServices.ComTypes
     ///     retrieving and storing data, the IDataObject interface specifies
     ///     methods for enumerating available formats and managing connections
     ///     to advisory sinks for handling change notifications.
-    /// </devdoc>
+    /// </summary>
     [CLSCompliant(false)]
     [ComImport()]
     [Guid("0000010E-0000-0000-C000-000000000046")]
     [InterfaceTypeAttribute(ComInterfaceType.InterfaceIsIUnknown)]
     public interface IDataObject {
 
-        /// <devdoc>
+        /// <summary>
         ///     Called by a data consumer to obtain data from a source data object.
         ///     The GetData method renders the data described in the specified FORMATETC
         ///     structure and transfers it through the specified STGMEDIUM structure.
         ///     The caller then assumes responsibility for releasing the STGMEDIUM structure.
-        /// </devdoc>
+        /// </summary>
         void GetData([In] ref FORMATETC format, out STGMEDIUM medium);
 
-        /// <devdoc>
+        /// <summary>
         ///     Called by a data consumer to obtain data from a source data object.
         ///     This method differs from the GetData method in that the caller must
         ///     allocate and free the specified storage medium.
-        /// </devdoc>
+        /// </summary>
         void GetDataHere([In] ref FORMATETC format, ref STGMEDIUM medium);
 
-        /// <devdoc>
+        /// <summary>
         ///     Determines whether the data object is capable of rendering the data
         ///     described in the FORMATETC structure. Objects attempting a paste or
         ///     drop operation can call this method before calling IDataObject::GetData
         ///     to get an indication of whether the operation may be successful.
-        /// </devdoc>
+        /// </summary>
         [PreserveSig]
         int QueryGetData([In] ref FORMATETC format);
 
-        /// <devdoc>
+        /// <summary>
         ///     Provides a standard FORMATETC structure that is logically equivalent to one that is more
         ///     complex. You use this method to determine whether two different
         ///     FORMATETC structures would return the same data, removing the need
         ///     for duplicate rendering.
-        /// </devdoc>
+        /// </summary>
         [PreserveSig]
         int GetCanonicalFormatEtc([In] ref FORMATETC formatIn, out FORMATETC formatOut);
 
-        /// <devdoc>
+        /// <summary>
         ///     Called by an object containing a data source to transfer data to
         ///     the object that implements this method.
-        /// </devdoc>
+        /// </summary>
         void SetData([In] ref FORMATETC formatIn, [In] ref STGMEDIUM medium, [MarshalAs(UnmanagedType.Bool)] bool release);
 
-        /// <devdoc>
+        /// <summary>
         ///     Creates an object for enumerating the FORMATETC structures for a
         ///     data object. These structures are used in calls to IDataObject::GetData
         ///     or IDataObject::SetData.
-        /// </devdoc>
+        /// </summary>
         IEnumFORMATETC EnumFormatEtc(DATADIR direction);
 
-        /// <devdoc>
+        /// <summary>
         ///     Called by an object supporting an advise sink to create a connection between
         ///     a data object and the advise sink. This enables the advise sink to be
         ///     notified of changes in the data of the object.
-        /// </devdoc>
+        /// </summary>
         [PreserveSig]
         int DAdvise([In] ref FORMATETC pFormatetc, ADVF advf, IAdviseSink adviseSink, out int connection);
 
-        /// <devdoc>
+        /// <summary>
         ///     Destroys a notification connection that had been previously set up.
-        /// </devdoc>
+        /// </summary>
         void DUnadvise(int connection);
 
-        /// <devdoc>
+        /// <summary>
         ///     Creates an object that can be used to enumerate the current advisory connections.
-        /// </devdoc>
+        /// </summary>
         [PreserveSig]
         int EnumDAdvise(out IEnumSTATDATA enumAdvise);
     }
