+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/ComTypes/STATDATA.cs
