+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/ComTypes/TYMED.cs
