+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/HandleCollector.cs
@@ -8,13 +8,13 @@ namespace System.Runtime.InteropServices
 {
     public sealed class HandleCollector
     {
-        private const int DeltaPercent = 10; // this is used for increasing the threshold. 
-        private int _initialThreshold;
+        // Used for increasing the threshold. 
+        private const int DeltaPercent = 10;
         private int _threshold;
         private int _handleCount;
 
-        private int[] _gc_counts = new int[3];
-        private int _gc_gen = 0;
+        private int[] _gcCounts = new int[3];
+        private int _gcGeneration = 0;
 
         public HandleCollector(string name, int initialThreshold) :
             this(name, initialThreshold, int.MaxValue)
@@ -25,21 +25,19 @@ namespace System.Runtime.InteropServices
         {
             if (initialThreshold < 0)
             {
-                throw new ArgumentOutOfRangeException(nameof(initialThreshold), SR.Arg_NeedNonNegNumRequired);
+                throw new ArgumentOutOfRangeException(nameof(initialThreshold), initialThreshold, SR.Arg_NeedNonNegNumRequired);
             }
-
             if (maximumThreshold < 0)
             {
-                throw new ArgumentOutOfRangeException(nameof(maximumThreshold), SR.Arg_NeedNonNegNumRequired);
+                throw new ArgumentOutOfRangeException(nameof(maximumThreshold), maximumThreshold, SR.Arg_NeedNonNegNumRequired);
             }
-
             if (initialThreshold > maximumThreshold)
             {
-                throw new ArgumentException(SR.Arg_InvalidThreshold);
+                throw new ArgumentException(SR.Arg_InvalidThreshold, nameof(initialThreshold));
             }
 
             Name = name ?? string.Empty;
-            _initialThreshold = initialThreshold;
+            InitialThreshold = initialThreshold;
             MaximumThreshold = maximumThreshold;
             _threshold = initialThreshold;
             _handleCount = 0;
@@ -47,7 +45,7 @@ namespace System.Runtime.InteropServices
 
         public int Count => _handleCount;
 
-        public int InitialThreshold => _initialThreshold;
+        public int InitialThreshold { get; }
 
         public int MaximumThreshold { get; }
 
@@ -55,7 +53,7 @@ namespace System.Runtime.InteropServices
 
         public void Add()
         {
-            int gen_collect = -1;
+            int collectionGeneration = -1;
             Interlocked.Increment(ref _handleCount);
             if (_handleCount < 0)
             {
@@ -67,26 +65,26 @@ namespace System.Runtime.InteropServices
                 lock (this)
                 {
                     _threshold = _handleCount + (_handleCount / DeltaPercent);
-                    gen_collect = _gc_gen;
-                    if (_gc_gen < 2)
+                    collectionGeneration = _gcGeneration;
+                    if (_gcGeneration < 2)
                     {
-                        _gc_gen++;
+                        _gcGeneration++;
                     }
                 }
             }
 
-            if ((gen_collect >= 0) &&
-                    ((gen_collect == 0) ||
-                    (_gc_counts[gen_collect] == GC.CollectionCount(gen_collect))))
+            if ((collectionGeneration >= 0) &&
+                    ((collectionGeneration == 0) ||
+                    (_gcCounts[collectionGeneration] == GC.CollectionCount(collectionGeneration))))
             {
-                GC.Collect(gen_collect);
-                Thread.Sleep(10 * gen_collect);
+                GC.Collect(collectionGeneration);
+                Thread.Sleep(10 * collectionGeneration);
             }
 
-            //don't bother with gen0. 
+            // Don't bother with gen0. 
             for (int i = 1; i < 3; i++)
             {
-                _gc_counts[i] = GC.CollectionCount(i);
+                _gcCounts[i] = GC.CollectionCount(i);
             }
         }
 
@@ -103,21 +101,21 @@ namespace System.Runtime.InteropServices
             {
                 lock (this)
                 {
-                    if (newThreshold > _initialThreshold)
+                    if (newThreshold > InitialThreshold)
                     {
                         _threshold = newThreshold;
                     }
                     else
                     {
-                        _threshold = _initialThreshold;
+                        _threshold = InitialThreshold;
                     }
-                    _gc_gen = 0;
+                    _gcGeneration = 0;
                 }
             }
 
             for (int i = 1; i < 3; i++)
             {
-                _gc_counts[i] = GC.CollectionCount(i);
+                _gcCounts[i] = GC.CollectionCount(i);
             }
         }
     }
