+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/SetWin32ContextInIDispatchAttribute.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [Obsolete("This attribute has been deprecated.  Application Domains no longer respect Activation Context boundaries in IDispatch calls.", error: false)]
+    [AttributeUsage(AttributeTargets.Assembly, Inherited = false)]
+    public sealed class SetWin32ContextInIDispatchAttribute : Attribute
+    {
+        public SetWin32ContextInIDispatchAttribute()
+        {
+        }
+    }
+}
