+++ src/System.Runtime.InteropServices/src/System/Runtime/InteropServices/TypeLibFuncAttribute.cs
@@ -1,0 +1,22 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [AttributeUsage(AttributeTargets.Method, Inherited = false)]
+    public sealed class TypeLibFuncAttribute : Attribute
+    {
+        public TypeLibFuncAttribute(TypeLibFuncFlags flags)
+        {
+            Value = flags;
+        }
+
+        public TypeLibFuncAttribute(short flags)
+        {
+            Value = (TypeLibFuncFlags)flags;
+        }
+
+        public TypeLibFuncFlags Value { get; }
+    }
+}
